Updated Bainbridge Files
These files are an update of my final project for the App Development course. They include the R-file for local launch of the shiny application (app.R), the associated data files used in the current version of the app (SlopeBio.csv, wracked.csv, and AllSed.csv), and finally the image of the sites for the project (bblayout.png).